min-max-heap
The program will promt the user to enter values to enter in a heap. Then then program will build the heap.
After building the heap the program will then find the minimum and maximum element in the heap.
Then the program will insert an element to the heap.
Then it will prompt the user if they would like to delete the minimum and maximum elements.